Add files via upload
Create a package and then add Cancellation.java code and Reservation.java code into it. pro.java is the main class that you should run in which you should import the created package. And provide the paths to the first, second, and third.txt files. Upon the user's decision the first, second, and third.txt file contents change.